+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Alert/AlertButtonProtocol.swift
@@ -21,5 +21,7 @@ public protocol AlertButtonProtocol: AlertButtonConvertible {
 }
 
 extension AlertButtonProtocol {
-    public func toButtons() -> [any AlertButtonProtocol] { [self] }
+    public func toButtons() -> [any AlertButtonProtocol] {
+        [self]
+    }
 }