+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Confirmation Dialog/ConfirmationDialogButtonProtocol.swift
@@ -21,5 +21,7 @@ public protocol ConfirmationDialogButtonProtocol: ConfirmationDialogButtonConver
 }
 
 extension ConfirmationDialogButtonProtocol {
-    public func toButtons() -> [any ConfirmationDialogButtonProtocol] { [self] }
+    public func toButtons() -> [any ConfirmationDialogButtonProtocol] {
+        [self]
+    }
 }