+++ Sources/VCore/Helpers/Architectural Pattern Helpers/UIKit/UI Action Sheet/UIActionSheetButtonProtocol.swift
@@ -18,7 +18,9 @@ public protocol UIActionSheetButtonProtocol: UIActionSheetButtonConvertible {
 }
 
 extension UIActionSheetButtonProtocol {
-    public func toButtons() -> [any UIActionSheetButtonProtocol] { [self] }
+    public func toButtons() -> [any UIActionSheetButtonProtocol] {
+        [self]
+    }
 }
 
 #endif